+++ R/distdescplot.R
@@ -13,10 +13,14 @@
 distdescplot <- function(x=x){
   x <- xts::as.xts(x[, 2], order.by = x$date)
   x.stationary <- x-mean(x)
-  fit <- fGarch::garchFit(~garch(1, 1), data = x.stationary, trace = FALSE)
-  if (xts::periodicity(x)$scale=="daily") {garchvol <- fit@sigma.t * sqrt(252)}
-  if (xts::periodicity(x)$scale=="weekly") {garchvol <- fit@sigma.t * sqrt(52)}
-  if (xts::periodicity(x)$scale=="monthly") {garchvol <- fit@sigma.t * sqrt(12)}
+  # fit <- fGarch::garchFit(~garch(1, 1), data = x.stationary, trace = FALSE)
+  # if (xts::periodicity(x)$scale=="daily") {garchvol <- fit@sigma.t * sqrt(252)}
+  # if (xts::periodicity(x)$scale=="weekly") {garchvol <- fit@sigma.t * sqrt(52)}
+  # if (xts::periodicity(x)$scale=="monthly") {garchvol <- fit@sigma.t * sqrt(12)}
+  fit <-  rugarch::ugarchfit(data = x, spec = rugarch::ugarchspec(), solver = "hybrid")
+  if (xts::periodicity(x)$scale=="daily") {garchvol <- fit@fit$sigma * sqrt(252)}
+  if (xts::periodicity(x)$scale=="weekly") {garchvol <- fit@fit$sigma * sqrt(52)}
+  if (xts::periodicity(x)$scale=="monthly") {garchvol <- fit@fit$sigma * sqrt(12)}
   voldata <- merge(x, garchvol)
   colnames(voldata) <- c("returns", "garch vol")
   voldata$CumulativeReturn <- cumsum(voldata$returns)