+++ R/globals.R
@@ -7,4 +7,6 @@ utils::globalVariables(c("df_fut","dflong","dfwide","expiry_table","holidaysOil"
                          "freq","remainder","week","u","sigma","res","s","e",
                          "calDays","hol","bizDays","futs","type","expirationDate",
                          "ticker","name","SupplyNOPEC","SupplyOPEC","Demand","Inv_Change",
-                         "tickers_eia","sd_category","tick.r","imports","exports","demand","supply"))
+                         "tickers_eia","sd_category","tick.r","imports","exports","demand","supply",
+                         "ref.opt.inputs","ref.opt.outputs","prices","LightSweet.yield","LightSweet",
+                         "HeavySour","HeavySour.yield","Element","product","max.prod","info"))