+++ NEWS.md
@@ -5,6 +5,7 @@
 + `chart_eia_steo()` returns a Supply/Demand balance from the EIA STEO data set. Currently configured for Global Liquids and will be augmented for US Crude, Light and Middle Distillates.
 + `chart_eia_sd()` returns Supply/Demand balance from the EIA weekly data for mogas, distillates, jet and resids.
 + `tickers_eia` table updated to build Supply Demand Balances for US products.
++ Datasets `ref.opt.inputs` and `ref.opt.outputs` to support refinery LP optimization education using `lpSolve` package. 
 
 ## Updates and Fixes
 + `eia2tidy()` fix for key variable in function.